Fix: get tasks not working

+++ TaskManagementSystem.Server/Models/Task.cs
@@ -6,7 +6,7 @@ public class TaskModel
   public required string Name { get; set; }
   public string? Description { get; set; }
   public long TaskListId { get; set; }
-  public IList<long> SubTaskIds { get; set; } = new List<long>();
+  public IList<int> SubTaskIds { get; set; } = new List<int>();
   public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
   public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 }
